Make Gravity Gloves pull more forgiving.
Instead of requiring the player to aim exactly at the object to pull,
we consider the set of objects near the crosshair and give precedence
to certain important items, as long as they are visible.

This makes the pull system a lot smoother, especially for controller
players, for whom aiming exactly at a small ammo clip or resin can be
cumbersome.